fix integer overflow

+++ handler/persistent_msg_relay_hdl/handler.go
@@ -143,5 +143,5 @@ func (h *Handler) createCleanupFile(msgIDs []string) error {
 
 func getDayHourUnix() int64 {
 	timeNow := time.Now().UTC()
-	return timeNow.Add(-time.Duration(int64(timeNow.Minute()*int(time.Minute) + timeNow.Second()*int(time.Second) + timeNow.Nanosecond()))).Unix()
+	return timeNow.Add(-time.Duration(int64(timeNow.Minute())*int64(time.Minute) + int64(timeNow.Second())*int64(time.Second) + int64(timeNow.Nanosecond()))).Unix()
 }
